+++ src/Localization/LOCALIZATIONS.mjs
@@ -1,3 +1,8 @@
+/** @typedef {import("./Localization.mjs").Localization} Localization */
+
+/**
+ * @type {Localization[]}
+ */
 export const LOCALIZATIONS = Object.freeze([
     {
         "fallback-languages": Object.freeze([
@@ -7,13 +12,11 @@ export const LOCALIZATIONS = Object.freeze([
             "de-LI"
         ]),
         getTexts: async () => (await import("./TEXTS_DE.mjs")).TEXTS_DE,
-        label: "Deutsch",
         language: "de"
     },
     {
         "fallback-default": true,
         getTexts: async () => (await import("./TEXTS_EN.mjs")).TEXTS_EN,
-        label: "English",
         language: "en"
     }
 ].map(localization => Object.freeze(localization)));