+++ src/Localization/Localization.mjs
@@ -1,3 +1,5 @@
+/** @typedef {import("../FluxLocalizationApi.mjs").FluxLocalizationApi} FluxLocalizationApi */
+
 /**
- * @typedef {{direction?: string | null, "fallback-default"?: boolean | null, "fallback-languages"?: string[] | null, getTexts: () => Promise<{[key: string]: string}>, label: string, language: string}} Localization
+ * @typedef {{direction?: string | null, "fallback-default"?: boolean | null, "fallback-languages"?: string[] | null, getLabel?: (localization: FluxLocalizationApi) => Promise<string>, getTexts?: () => Promise<{[key: string]: string}>, language: string}} Localization
  */